(#181) organizando metodos do user
