+++ app/ext/auth.py
@@ -1,5 +1,5 @@
 from flask_jwt_extended import JWTManager, verify_jwt_in_request, get_jwt
-from ..model import user
+from ..services import user_services
 from functools import wraps
 
 jwt = JWTManager()
@@ -33,4 +33,4 @@ def user_identity_lookup(user):
 @jwt.user_lookup_loader
 def user_lookup_callback(_jwt_header, jwt_data):
     identity = jwt_data["sub"]
-    return user.get(reg=identity)
+    return user_services.get(reg=identity)
